Move sensors to interfaces
